requirements.txt
kafka-python==2.0.2
torch==2.1.2